requirements.txt
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.7
Django==3.2.9
django-bootstrap-form==3.4
django-crispy-forms==1.13.0
django-filter==21.1
idna==3.3
PyJWT==2.3.0
pytz==2021.3
requests==2.26.0
sqlparse==0.4.2
twilio==7.3.0
urllib3==1.26.7
